Refactor: libpacemaker: move ordering functions to own source file
The ordering constraint functions from pcmk_sched_constraints.c are now in a
new file pcmk_sched_ordering.c. There are no code changes aside from formatting
and comments.

At this point, pcmk_sched_constraints.c (which went from nearly 1,500 lines to
under 400) contains only functions for constraints in general, and not anything
specific to one type.
